refactor filesystem reads out of index into standalone module
